Convert to flake8
I went with pycodestyle initially due to a talk about the pep8 utility becoming
pycodestyle. Flake8 however still has a few other checks that are not part of
pycodestyle so having those in there earlier is better than later. Once flake8
upgrades to pycodestyle it should basically be transparent to us anyhow.